TPMModule: fixed quote() and quote2() functions of NIARL_TPM_Module
The above functions display the values of PCRs to be added to a TPM quote
before performing the quote operation itself. This solution is not correct
because there could be a PCR extend operation between the PCR read and
the quote operation. If this condition happens, the report validation
done by the Appraiser will fail.

This patch overcomes the problem by reading the PCR values from the PCR
Composite TSS object after the quote as, according to TCG specifications,
the object is filled with correct values by the Tspi_TPM_Quote() and
Tspi_TPM_Quote2() functions.

Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>